feat: display the change details in a dedicated view
