+++ opencve/admin.py
@@ -1,6 +1,4 @@
 import datetime
-import json
-from difflib import HtmlDiff
 
 import arrow
 from flask import abort
@@ -10,29 +8,6 @@ from flask_user import current_user
 from sqlalchemy import desc, func
 from sqlalchemy.orm import joinedload
 from wtforms import PasswordField, validators
-
-
-class CustomHtmlHTML(HtmlDiff):
-    def __init__(self, *args, **kwargs):
-        self._table_template = """
-        <table class="table table-condensed">
-            <thead>
-                <tr>
-                    <th colspan="2">Old JSON</th>
-                    <th colspan="2">New JSON</th>
-                </tr>
-            </thead>
-            <tbody>%(data_rows)s</tbody>
-        </table>"""
-        super().__init__(*args, **kwargs)
-
-    def _format_line(self, side, flag, linenum, text):
-        text = text.replace("&", "&amp;").replace(">", "&gt;").replace("<", "&lt;")
-        text = text.replace(" ", "&nbsp;").rstrip()
-        return '<td class="diff_header">%s</td><td class="break">%s</td>' % (
-            linenum,
-            text,
-        )
 
 
 class AuthModelView(ModelView):
@@ -213,34 +188,6 @@ class HomeView(AdminIndexView):
         )
 
         return self.render("admin/task.html", task=task, changes=changes)
-
-    @expose("/changes/<id>")
-    def change(self, id):
-        from .models.changes import Change
-
-        change = Change.query.get(id)
-        previous = (
-            Change.query.filter(Change.created_at < change.created_at)
-            .filter(Change.cve == change.cve)
-            .order_by(Change.created_at.desc())
-            .first()
-        )
-
-        if previous:
-            previous_json = previous.json
-        else:
-            previous_json = {}
-
-        differ = CustomHtmlHTML()
-        diff = differ.make_table(
-            fromlines=json.dumps(previous_json, sort_keys=True, indent=2).split("\n"),
-            tolines=json.dumps(change.json, sort_keys=True, indent=2).split("\n"),
-            context=True,
-        )
-
-        return self.render(
-            "/admin/change.html", change=change, previous=previous, diff=diff
-        )
 
 
 class UserModelView(AuthModelView):
