+++ opencve/utils.py
@@ -1,4 +1,5 @@
 from nested_lookup import nested_lookup
+from difflib import HtmlDiff
 
 from opencve.constants import PRODUCT_SEPARATOR
 from opencve.models.cwe import Cwe
@@ -55,3 +56,26 @@ def get_cwes_details(problems):
         if cwe:
             cwes[cwe_id] = cwe.name
     return cwes
+
+
+class CustomHtmlHTML(HtmlDiff):
+    def __init__(self, *args, **kwargs):
+        self._table_template = """
+        <table class="table table-diff table-condensed">
+            <thead>
+                <tr>
+                    <th colspan="2">Old JSON</th>
+                    <th colspan="2">New JSON</th>
+                </tr>
+            </thead>
+            <tbody>%(data_rows)s</tbody>
+        </table>"""
+        super().__init__(*args, **kwargs)
+
+    def _format_line(self, side, flag, linenum, text):
+        text = text.replace("&", "&amp;").replace(">", "&gt;").replace("<", "&lt;")
+        text = text.replace(" ", "&nbsp;").rstrip()
+        return '<td class="diff_header">%s</td><td class="break">%s</td>' % (
+            linenum,
+            text,
+        )
